@@ -1,82 +1,157 @@
 package com.dezen.riccardo.musicplayer;
 
-import android.media.MediaPlayer;
+import android.content.ComponentName;
+import android.content.Context;
+import android.content.Intent;
+import android.content.ServiceConnection;
 import android.os.Bundle;
-import android.provider.MediaStore;
+import android.os.IBinder;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ListView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
-import androidx.loader.app.LoaderManager;
-import androidx.loader.content.Loader;
 
-import java.io.IOException;
-import java.util.List;
+import com.dezen.riccardo.musicplayer.song.SongManager;
 
-public class SongListFragment extends Fragment{
+/**
+ * Fragment displaying the song list for the app.
+ *
+ * @author Riccardo De Zen.
+ */
+public class SongListFragment extends Fragment {
 
-    private SongViewModel songList;
+    private MusicController musicController;
+    private SongManager songManager;
     private ListView songsListView;
-    private MediaPlayer mediaPlayer;
+    private View rootView;
 
+    private int currentSong;
+
+    /**
+     * Defines callbacks for service binding, passed to bindService()
+     */
+    private ServiceConnection serviceConnection = new ServiceConnection() {
+        /**
+         * @param className The class name of the service.
+         * @param service The returned Binder.
+         */
+        @Override
+        public void onServiceConnected(ComponentName className, IBinder service) {
+            // We've bound to LocalService, cast the IBinder and get LocalService instance
+            musicController = (MusicController) service;
+            musicController.play(currentSong);
+        }
+
+        /**
+         * @param arg0 The component that disconnected.
+         */
+        @Override
+        public void onServiceDisconnected(ComponentName arg0) {
+            musicController = null;
+        }
+    };
+
+    /**
+     * Initializes the {@link SongManager}.
+     *
+     * @param context The Context that hosts the Fragment.
+     */
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        songList = ViewModelProviders.of(this).get(SongViewModel.class);
-        songList.getSongList().observe(this, new Observer<List<Song>>() {
-            @Override
-            public void onChanged(List<Song> songs) {
-                songsListView.setAdapter(new CustomAdapter());
-            }
-        });
-        songsListView = getView().findViewById(R.id.songs_listview);
-        songsListView.setAdapter(new CustomAdapter());
-        mediaPlayer = new MediaPlayer();
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+        songManager = SongManager.getInstance(context);
     }
 
+    /**
+     * Inflates the View for this Fragment.
+     *
+     * @param inflater           The inflater to use.
+     * @param container          The parent container.
+     * @param savedInstanceState The saved instance state if any.
+     * @return An inflated {@link R.layout#fragment_songlist} view.
+     */
     @Nullable
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_songlist, container, false);
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        rootView = inflater.inflate(R.layout.fragment_songlist, container, false);
+        return rootView;
     }
 
-    private void play(String path){
-        if(mediaPlayer.isPlaying()) mediaPlayer.stop();
-        mediaPlayer.reset();
-        try{
-            mediaPlayer.setDataSource(path);
-            mediaPlayer.prepare();
-            mediaPlayer.start();
-        }catch(IOException e){
-            Toast.makeText(getContext(), "Impossibile riprodurre il file", Toast.LENGTH_SHORT).show();
+    /**
+     * Sets the list's adapter to listen to the song list.
+     */
+    @Override
+    public void onStart() {
+        super.onStart();
+        songManager.getSongs().observe(this,
+                songs -> songsListView.setAdapter(new CustomAdapter()));
+        songsListView = rootView.findViewById(R.id.songs_listview);
+        songsListView.setAdapter(new CustomAdapter());
+
+        if (musicController != null)
+            musicController.resume();
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        if (musicController != null)
+            musicController.pause();
+    }
+
+    @Override
+    public void onDestroy() {
+        super.onDestroy();
+        if (getActivity() != null)
+            getActivity().unbindService(serviceConnection);
+    }
+
+    /**
+     * This method attempts to bind to the {@link PlayerService} class to play a song.
+     * The method attempts the binding only if the IBinder is {@code null}.
+     *
+     * @param position The song to play on the Service.
+     */
+    private void play(int position) {
+        if (getActivity() == null)
+            return;
+
+        currentSong = position;
+
+        if (musicController != null)
+            musicController.play(position);
+        else {
+            // Making sure the service is started.
+            getActivity().startService(new Intent(getContext(), PlayerService.class));
+            // Binding to the service.
+            getActivity().bindService(
+                    new Intent(getContext(), PlayerService.class),
+                    serviceConnection,
+                    Context.BIND_AUTO_CREATE
+            );
         }
     }
 
-    private View getItemView(final int index){
-        final View newView = getLayoutInflater().inflate(R.layout.song_listview_item, null);
-        ((TextView)newView.findViewById(R.id.textView_song_title)).setText(songList.get(index).getTitle());
-        newView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //play(songList.get(index).getData());
-            }
-        });
+    private View getItemView(final int index) {
+        View newView = getLayoutInflater().inflate(R.layout.song_listview_item, null);
+        ((TextView) newView.findViewById(R.id.textView_song_title)).setText(songManager.getSongs().getValue().get(index).getTitle());
+        newView.setOnClickListener(v ->
+                play(index)
+        );
         return newView;
     }
 
     private class CustomAdapter extends BaseAdapter {
         @Override
         public int getCount() {
-            return songList.size();
+            return songManager.getSongs().getValue().size();
         }
 
         @Override
